LayoutTests/fast/selectors/lang-invalid-expected.txt

Verify invalid :lang() selectors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll(":lang()").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(lang()").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(:lang(id)").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(:lang(en, br)").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..

PASS document.querySelectorAll(":lang(<0_0>)").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(9)").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(99)").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(999)").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..

PASS document.querySelectorAll(":lang(})").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang({)").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang({}")).length threw exception SyntaxError: Unexpected token ')'.
PASS document.querySelectorAll(":lang(()").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang())").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(()")).length threw exception SyntaxError: Unexpected token ')'.
PASS document.querySelectorAll(":lang(])").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang([)").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang([])").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(@media screen {})").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(@font-face {})").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(:)").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(;)").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(.)").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(+)").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(-)").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(=)").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS document.querySelectorAll(":lang(&&)").length threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..

PASS getComputedStyle(document.getElementById("target1")).color is "rgb(0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
